+++ src/main/java/org/psychohelp/psychohelp/repository/CreneauRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CreneauRepository extends JpaRepository<Creneau, Long> {
+public interface CreneauRepository extends JpaRepository<Creneau, Integer> {
 
     List<Creneau> findByStatutTrue();
